Split tests by file
