+++ __tests__/main.test.ts
@@ -1,46 +1,7 @@
-import * as config from '../src/config'
+import {run} from '../src/main'
 
-test('No API Token', async () => {
-  expect(() => config.create_config()).toThrow(
+test('Empty Test', async () => {
+  expect(run()).rejects.toThrow(
     'Need to have either an api_token or global_token with email set'
   )
-})
-
-test('No Zone ID', async () => {
-  process.env['INPUT_API_TOKEN'] = '2'
-  expect(() => config.create_config()).toThrow(
-    'Input required and not supplied: zone'
-  )
-})
-
-test('Legacy Auth', async () => {
-  process.env['INPUT_ZONE'] = '2'
-  process.env['INPUT_API_TOKEN'] = ''
-  process.env['INPUT_GLOBAL_TOKEN'] = '4'
-  process.env['INPUT_EMAIL'] = 'cyberjake@pm.me'
-  let conf = config.create_config()
-  expect(conf.token_method).toEqual('legacy')
-  expect((conf.instance.defaults.headers['X-Auth-Key'] = '4'))
-  expect((conf.instance.defaults.headers['X-Auth-Email'] = 'cyberjake@pm.me'))
-})
-
-test('Purge Everything', async () => {
-  process.env['INPUT_API_TOKEN'] = '1'
-  process.env['INPUT_ZONE'] = '2'
-  let conf = config.create_config()
-  expect(conf.token_method).toEqual('token')
-  expect(conf.zone_id).toEqual('2')
-  expect(conf.instance.defaults.headers['Authorization']).toEqual('Bearer 1')
-  expect(conf.purge_body).toEqual({purge_everything: true})
-})
-
-test('Purge URLs Config', async () => {
-  process.env['INPUT_API_TOKEN'] = '1'
-  process.env['INPUT_ZONE'] = '2'
-  process.env['INPUT_URLS'] = 'url1\nurl2'
-  let conf = config.create_config()
-  expect(conf.token_method).toEqual('token')
-  expect(conf.zone_id).toEqual('2')
-  expect(conf.instance.defaults.headers['Authorization']).toEqual('Bearer 1')
-  expect(conf.purge_body).toEqual({files: ['url1', 'url2']})
 })
