+++ package.json
@@ -9,7 +9,7 @@
     "format": "prettier --write **/*.ts",
     "format-check": "prettier --check **/*.ts",
     "lint": "eslint src/**/*.ts",
-    "test": "jest",
+    "test": "jest --coverage",
     "all": "yarn run format-check && yarn run lint && yarn run test"
   },
   "repository": {
